Rename sparseblock into blocksparse. Move tests to specific file.
