+++ theano/sandbox/blocksparse.py
@@ -1,9 +1,11 @@
-# TODO: imports :)
+import numpy
+import theano
+from theano import Op, Apply
+import theano.tensor as T
+from theano.tensor import discrete_dtypes
+
 
 class SparseBlockGemv(Op):
-    
-    register_opt = defaultdict(list)
-
     """
     This op computes the dot product of specified pieces of vectors
     and matrices, returning pieces of vectors.
@@ -17,6 +19,8 @@ class SparseBlockGemv(Op):
 
 
     """
+    __props__ = ('inplace',)
+
     def __init__(self, inplace=False):
         self.inplace = False
 
@@ -74,141 +78,69 @@ class SparseBlockGemv(Op):
     def perform(self, node, inp, out_):
         raise NotImplementedError('Optimization of SparseBlockGemv failed.')
 
-    def grad(self, inputs, grads):
-        o, W, h, inputIdx, outputIdx = inputs
-        go = grads[0]
-
-        Wgrad = sparse_block_outer(W.zeros_like(),
-                                      h, go, inputIdx, outputIdx)
-        hgrad = sparse_block_gemv(h.zeros_like(),
-                                     W.dimshuffle((1, 0, 3, 2)),
-                                     go,
-                                     outputIdx, inputIdx)
-        return [go, Wgrad, hgrad,
-                grad_undefined(self, 3, inputIdx,
-                               "grad of inputIdx makes no sense"),
-                grad_undefined(self, 4, outputIdx,
-                               "grad of outputIdx makes no sense")]
-
-                
-sparse_block_gemv = SparseBlockGemv(False)
-sparse_block_gemv_inplace = SparseBlockGemv(True)
-
-
-class SparseBlockOuter(Op):
-    """
-    This computes the outer product of two sets of pieces of vectors
-    updating a full matrix with the results.
-
-    It computes something like this:
-
-      o[i, j] = (alpha * outer(x[i], y[j])) + o[i, j]
-
-    The i and j are taken from the xIdx and yIdx lists respectively.
-
-    This op should not be called directly since its interface is
-    subject to change without notice.  It is involved in the gradient
-    of SparseBlockGemvSS.
-    """
-    def __init__(self, inplace=False):
-        self.inplace = inplace
-        if self.inplace:
-            self.destroy_map = {0: [0]}
-
-    def make_node(self, o, x, y, xIdx, yIdx, alpha=None):
-        """
-            TODO: WRITEME
-        """
-        one = tensor.constant(numpy.asarray(1.0, dtype='float32'))
-        o = basic_ops.as_cuda_ndarray_variable(o)
-        x = basic_ops.as_cuda_ndarray_variable(x)
-        y = basic_ops.as_cuda_ndarray_variable(y)
-        if alpha is None:
-            alpha = one
-        return Apply(self, [o, x, y, xIdx, yIdx, alpha],
-                     [o.type()])
-
-    def perform(self, node, inp, out_):
-        raise NotImplementedError('Optimization of SparseBlockOuter failed.')
-
     def grad(self, inputs, output_gradients):
-        meta_grad_op = MetaGradSparseBlockGemv(output_gradients)
-        return [meta_grad_op(inp) for inp in inputs]
+        # meta_grad_op = MetaGradSparseBlockGemv(output_gradients)
+        # return [meta_grad_op(inp) for inp in inputs]
+        return # TODO
 
 
-sparse_block_outer = SparseBlockOuter(False)
-sparse_block_outer_inplace = SparseBlockOuter(True)
-
-
-def cpu_sparse_block_gemv(o, W, h, inputIdx, outputIdx):
+def sparse_block_gemv_cpu(W, h, inputIdx, bias, outputIdx):
     """
     Creates a graph for the sparse block dot operation. Check SparseBlockGemv's
     docstring for information about the arguments.
     """
-    def _outer_loop_over_outputIdx(i, W, h, inputIdx, outputIdx):
 
-        def _inner_loop_over_inputIdx(j, W, h, inputIdx, outputIdx):
+    def _loop_over_batch(b):
 
-            return T.dot(h[:, j, :], W[inputIdx[j], outputIdx[i], outputIdx[i], :, :])
+        def _loop_over_outputIdx(i):
 
-        a = theano.scan(fn=_inner_loop_over_inputIdx,
-                        sequences=T.arange(0, inputIdx.shape[1]),
-                        non_sequences=[W, h, inputIdx, outputIdx],
-                        name='_inner_loop_over_inputIdx')[0]
-        return a.sum()
+            def _loop_over_inputIdx(j):
+                return T.dot(h[b, j, :], W[inputIdx[b, j], outputIdx[b, i], :, :])
 
-    b = theano.scan(fn=_outer_loop_over_outputIdx,
-                    sequences=T.arange(0, outputIdx.shape[1]),
-                    non_sequences=[W, h, inputIdx, outputIdx],
-                    name='_outer_loop_over_outputIdx')[0]
+            res3 = theano.scan(fn=_loop_over_inputIdx,
+                            sequences=T.arange(0, inputIdx.shape[1]),
+                            name='_loop_over_inputIdx')[0]
 
-    return T.set_subtensor(o[:, :, :], b.dimshuffle(1, 0, 2))
+            return res3.sum(axis=0)
 
+        res2 = theano.scan(fn=_loop_over_outputIdx,
+                        sequences=T.arange(0, outputIdx.shape[1]),
+                        name='_loop_over_outputIdx')[0]
 
-def cpu_sparse_block_outer(o, x, y, xIdx, yIdx, alpha=None):
-    """
-        TODO: WRITEME
-    """
+        return res2
 
-    # TODO 
+    res1 = theano.scan(fn=_loop_over_batch,
+                       sequences=T.arange(0, inputIdx.shape[0]),
+                       name='_loop_over_batch')[0]
 
-    return None
+    return res1 + bias.take(outputIdx, axis=0)
 
 
-def sparse_block_dot(W, h, inputIdx, b, outputIdx, inplace=False):
-    """
-    Compute the dot product (plus bias) of the specified pieces of vectors
-    and matrices.
+def sparse_block_outer_cpu(x, y, xIdx, yIdx, alpha=None):
+    if alpha is None:
+        alpha = T.constant(numpy.asarray(1.0, dtype='float32'))
 
-    Parameters
-    ----------
-    var: shape, comment
-    W: (iBlocks, oBlocks, iSize, oSize), weight matrix
-    h: (batch, iWin, iSize), input from lower layer (sparse)
-    inputIdx: (batch, iWin), indexes of the input blocks
-    b: (oBlocks, oSize), bias vector
-    outputIdx: (batch, oWin), indexes of the output blocks
+    def _loop_over_batch(b):
 
-    returns (batch, oWin, oSize), dot(W[i, j], h[i]) + b[j]
-         but b[j] is only added once
+        def _loop_over_outputIdx(i):
 
-    Notation
-    --------
-    - `batch` is the number of examples in a minibatch (batch size).
-    - `iBlocks` is the total number of blocks in the input (from lower layer).
-    - `iSize` is the size of each of these input blocks.
-    - `iWin` is the number of blocks that will be used as inputs. Which blocks
-      will be used is specified in `inputIdx`.
-    - `oBlocks` is the number or possible output blocks.
-    - `oSize` is the size of each of these output blocks.
-    - `oWin` is the number of output blocks that will actually be computed.
-      Which blocks will be computed is specified in `outputIdx`.
-    """
-    assert inputIdx.ndim == h.ndim - 1
-    assert outputIdx.ndim == inputIdx.ndim
-    if h.ndim == 2:
-        h = h.dimshuffle('x', 0, 1)
-        inputIdx = inputIdx.dimshuffle('x', 0)
-        outputIdx = outputIdx.dimshuffle('x', 0)
-    return SparseBlockGemv(inplace)(b.take(outputIdx, axis=0), W, h,
-                                inputIdx, outputIdx)
+            def _loop_over_inputIdx(j):
+                return T.outer(x[b, i, :], y[b, j, :])
+
+            res3 = theano.scan(fn=_loop_over_inputIdx,
+                            sequences=T.arange(0, xIdx.shape[1]),
+                            name='_loop_over_inputIdx')[0]
+
+            return res3
+
+        res2 = theano.scan(fn=_loop_over_outputIdx,
+                        sequences=T.arange(0, yIdx.shape[1]),
+                        name='_loop_over_outputIdx')[0]
+
+        return res2
+
+    res1 = theano.scan(fn=_loop_over_batch,
+                       sequences=T.arange(0, xIdx.shape[0]),
+                       name='_loop_over_batch')[0]
+
+    return alpha * res1.mean(axis=0)
